extract border utils
