+++ src/assets/theme/helpers/index.ts
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-//export * from './font-utils';
+export * from 'src/assets/theme/helpers/border-utils';
