+++ src/assets/theme/index.ts
@@ -37,14 +37,7 @@ import { createTheme, Theme } from '@mui/material';
 import { responsiveFontSizes } from './responsiveFont';
 
 // Material Kit 2 React base styles
-import {
-  palette,
-  breakpoints,
-  typography,
-  boxShadows,
-  borders,
-  globals,
-} from './base';
+import { palette, breakpoints, typography, boxShadows, globals } from './base';
 
 // Material Kit 2 React helper functions
 import {
@@ -93,7 +86,6 @@ export const theme: Theme = {
   ...responsiveFontSizes(generatedTheme, {
     breakpoints: generatedTheme.breakpoints.keys,
   }),
-  borders,
   boxShadows,
   functions: {
     boxShadow,
